requirements.txt
pysmb>=1.2.9 